+++ internal/graph/check.go
@@ -31,7 +31,6 @@ var tracer = otel.Tracer("internal/graph/check")
 type setOperatorType int
 
 var (
-	ErrShortCircuit       = errors.New("short circuit")
 	ErrUnknownSetOperator = fmt.Errorf("%w: unexpected set operator type encountered", openfgaErrors.ErrUnknown)
 )
 
@@ -1153,6 +1152,9 @@ func (c *LocalChecker) checkDirect(parentctx context.Context, req *ResolveCheckR
 				if typesys.UsersetCanFastPathWeight2(objectType, relation, userType, directlyRelatedUsersetTypes) {
 					resolver = c.checkUsersetFastPathV2
 					span.SetAttributes(attribute.String("resolver", "fastpathv2"))
+				} else if typesys.RecursiveUsersetCanFastPathV2(objectType, relation, userType) {
+					resolver = c.recursiveUsersetFastPathV2
+					span.SetAttributes(attribute.String("resolver", "recursivefastpathv2"))
 				}
 			} else if !isUserset {
 				if typesys.UsersetCanFastPath(directlyRelatedUsersetTypes) {
@@ -1348,7 +1350,7 @@ func (c *LocalChecker) checkTTU(parentctx context.Context, req *ResolveCheckRequ
 		object := tk.GetObject()
 
 		span.SetAttributes(
-			attribute.String("tupleset_relation", fmt.Sprintf("%s#%s", tuple.GetType(object), tuplesetRelation)),
+			attribute.String("tupleset_relation", tuple.ToObjectRelationString(tuple.GetType(object), tuplesetRelation)),
 			attribute.String("computed_relation", computedRelation),
 		)
 
@@ -1392,9 +1394,6 @@ func (c *LocalChecker) checkTTU(parentctx context.Context, req *ResolveCheckRequ
 				resolver = c.checkTTUFastPathV2
 				span.SetAttributes(attribute.String("resolver", "fastpathv2"))
 			} else if typesys.RecursiveTTUCanFastPathV2(objectType, relation, userType, rewrite.GetTupleToUserset()) {
-				// less common
-				// TODO when this "if" is taken out of the optimization flag, we can remove RecursiveTTUCanFastPath,
-				// since this code is a generalization of it.
 				resolver = c.recursiveTTUFastPathV2
 				span.SetAttributes(attribute.String("resolver", "recursivefastpathv2"))
 			}