+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Try to keep listed changes to a concise bulleted list of simple explanations of changes. Aim for the amount of information needed so that readers can understand where they would look in the codebase to investigate the changes' implementation, or where they would look in the documentation to understand how to make use of the change in practice - better yet, link directly to the docs and provide detailed information there. Only elaborate if doing so is required to avoid breaking changes or experimental features from ruining someone's day.
 
 ## [Unreleased]
+### Added
+- Added `storage.ErrTransactionThrottled` for throttling errors applied at the datastore level. [#2304](https://github.com/openfga/openfga/pull/2304).
+
 ### Changed
 - Rewrite some of the check resolution logic to keep track of running go routines and ensure that all routines have synchronized back into the original parent thread of execution before continuing. [#2275](https://github.com/openfga/openfga/pull/2275)
+
+### Removed
+- Removed recently-added `tuples_iterator_cache_invalid_hit_count` metric. The `cachecontroller_cache_invalidation_count` from 1.8.6 better accomplishes the same goal. [#2296)[https://github.com/openfga/openfga/pull/2296/]
 
 ### Fixed
 - Fixed evaluation of certain recursive TTU cases behind the `enable-check-optimizations` flag. [#2281](https://github.com/openfga/openfga/pull/2281)